Conflicts:
	.idea/modules.xml
	oap-application/pom.xml
	oap-cli/pom.xml
	oap-etl/pom.xml
	oap-http/pom.xml
	oap-json/pom.xml
	oap-logstream/pom.xml
	oap-metrics/pom.xml
	oap-stdlib/pom.xml
	oap-stdlib/src/main/java/oap/io/Files.java
	oap-storage/pom.xml
	oap-ws/pom.xml
	pom.xml

+++ oap-logstream/src/main/java/oap/logstream/LogWriter.java
@@ -45,12 +45,12 @@ public class LogWriter implements Closeable {
     private final String prefix;
     private final String root;
     private int bufferSize;
-    private int interval;
+    private long interval;
     private OutputStream out;
     private String lastPattern;
     private Scheduled scheduled;
 
-    public LogWriter( String prefix, String root, String suffix, int bufferSize, int interval ) {
+    public LogWriter( String prefix, String root, String suffix, int bufferSize, long interval ) {
         this.prefix = prefix;
         this.root = root;
         this.suffix = suffix;
